requirements.txt
flask
python-magic
python-dotenv
